rtdp CMakeLists.txt and demo run log
