+++ .gitignore
@@ -1,7 +1,13 @@
-# Python cache files
+# Python files
 __pycache__
 .DS_Store
-
+.venv/
+env/
+venv/
+*.pyc
+*.pyo
+*.pyd
+*.pdb
 # Java Binary files
 *.class
 *.jar
@@ -29,10 +35,6 @@ prom-data/
 influxdb-data/
 influxdb-config/
 grafana-data/
-
-# SQLite3 DB files
-*.db
-.nfs*
 
 # Add output directory with large binary files
 output/
