+++ rtdp/demo_run_log/sink_ejfat-6.txt
@@ -1,0 +1,61 @@
+[2025-09-26 02:35:22.445 UTC]  Incoming: [0.043 Gbps] (5.402 MB/s)  total=10.81344 MB
+[2025-09-26 02:35:24.447 UTC]  Incoming: [0.073 Gbps] (9.181 MB/s)  total=29.196288 MB
+[2025-09-26 02:35:26.449 UTC]  Incoming: [0.073 Gbps] (9.182 MB/s)  total=47.579136 MB
+[2025-09-26 02:35:28.451 UTC]  Incoming: [0.069 Gbps] (8.642 MB/s)  total=64.88064 MB
+[2025-09-26 02:35:30.453 UTC]  Incoming: [0.073 Gbps] (9.182 MB/s)  total=83.263488 MB
+[2025-09-26 02:35:32.455 UTC]  Incoming: [0.069 Gbps] (8.643 MB/s)  total=100.564992 MB
+[2025-09-26 02:35:34.457 UTC]  Incoming: [0.073 Gbps] (9.182 MB/s)  total=118.94784 MB
+[2025-09-26 02:35:36.459 UTC]  Incoming: [0.069 Gbps] (8.643 MB/s)  total=136.249344 MB
+[2025-09-26 02:35:38.461 UTC]  Incoming: [0.073 Gbps] (9.182 MB/s)  total=154.632192 MB
+[2025-09-26 02:35:40.463 UTC]  Incoming: [0.069 Gbps] (8.643 MB/s)  total=171.933696 MB
+[2025-09-26 02:35:42.465 UTC]  Incoming: [0.073 Gbps] (9.182 MB/s)  total=190.316544 MB
+[2025-09-26 02:35:44.468 UTC]  Incoming: [0.073 Gbps] (9.182 MB/s)  total=208.699392 MB
+[2025-09-26 02:35:46.470 UTC]  Incoming: [0.069 Gbps] (8.642 MB/s)  total=226.000896 MB
+[2025-09-26 02:35:48.472 UTC]  Incoming: [0.069 Gbps] (8.642 MB/s)  total=243.3024 MB
+[2025-09-26 02:35:50.474 UTC]  Incoming: [0.069 Gbps] (8.642 MB/s)  total=260.603904 MB
+[2025-09-26 02:35:52.476 UTC]  Incoming: [0.073 Gbps] (9.182 MB/s)  total=278.986752 MB
+[2025-09-26 02:35:54.478 UTC]  Incoming: [0.069 Gbps] (8.641 MB/s)  total=296.288256 MB
+[2025-09-26 02:35:56.480 UTC]  Incoming: [0.073 Gbps] (9.183 MB/s)  total=314.671104 MB
+[2025-09-26 02:35:58.482 UTC]  Incoming: [0.069 Gbps] (8.642 MB/s)  total=331.972608 MB
+[2025-09-26 02:36:00.484 UTC]  Incoming: [0.073 Gbps] (9.182 MB/s)  total=350.355456 MB
+[2025-09-26 02:36:02.486 UTC]  Incoming: [0.069 Gbps] (8.642 MB/s)  total=367.65696 MB
+[2025-09-26 02:36:04.488 UTC]  Incoming: [0.073 Gbps] (9.181 MB/s)  total=386.039808 MB
+[2025-09-26 02:36:06.490 UTC]  Incoming: [0.073 Gbps] (9.182 MB/s)  total=404.422656 MB
+[2025-09-26 02:36:08.492 UTC]  Incoming: [0.069 Gbps] (8.642 MB/s)  total=421.72416 MB
+[2025-09-26 02:36:10.494 UTC]  Incoming: [0.073 Gbps] (9.182 MB/s)  total=440.107008 MB
+[2025-09-26 02:36:12.496 UTC]  Incoming: [0.069 Gbps] (8.642 MB/s)  total=457.408512 MB
+[2025-09-26 02:36:14.499 UTC]  Incoming: [0.073 Gbps] (9.182 MB/s)  total=475.79136 MB
+[2025-09-26 02:36:16.501 UTC]  Incoming: [0.069 Gbps] (8.642 MB/s)  total=493.092864 MB
+[2025-09-26 02:36:18.503 UTC]  Incoming: [0.073 Gbps] (9.182 MB/s)  total=511.475712 MB
+[2025-09-26 02:36:20.505 UTC]  Incoming: [0.069 Gbps] (8.643 MB/s)  total=528.777216 MB
+[2025-09-26 02:36:22.507 UTC]  Incoming: [0.073 Gbps] (9.182 MB/s)  total=547.160064 MB
+[2025-09-26 02:36:24.509 UTC]  Incoming: [0.069 Gbps] (8.642 MB/s)  total=564.461568 MB
+[2025-09-26 02:36:26.511 UTC]  Incoming: [0.073 Gbps] (9.182 MB/s)  total=582.844416 MB
+[2025-09-26 02:36:28.513 UTC]  Incoming: [0.069 Gbps] (8.642 MB/s)  total=600.14592 MB
+[2025-09-26 02:36:30.515 UTC]  Incoming: [0.073 Gbps] (9.182 MB/s)  total=618.528768 MB
+[2025-09-26 02:36:32.517 UTC]  Incoming: [0.069 Gbps] (8.643 MB/s)  total=635.830272 MB
+[2025-09-26 02:36:34.519 UTC]  Incoming: [0.073 Gbps] (9.182 MB/s)  total=654.21312 MB
+[2025-09-26 02:36:36.521 UTC]  Incoming: [0.073 Gbps] (9.183 MB/s)  total=672.595968 MB
+[2025-09-26 02:36:38.523 UTC]  Incoming: [0.073 Gbps] (9.182 MB/s)  total=690.978816 MB
+[2025-09-26 02:36:40.525 UTC]  Incoming: [0.073 Gbps] (9.183 MB/s)  total=709.361664 MB
+[2025-09-26 02:36:42.527 UTC]  Incoming: [0.069 Gbps] (8.642 MB/s)  total=726.663168 MB
+[2025-09-26 02:36:44.529 UTC]  Incoming: [0.073 Gbps] (9.182 MB/s)  total=745.046016 MB
+[2025-09-26 02:36:46.531 UTC]  Incoming: [0.069 Gbps] (8.641 MB/s)  total=762.34752 MB
+[2025-09-26 02:36:48.533 UTC]  Incoming: [0.073 Gbps] (9.182 MB/s)  total=780.730368 MB
+[2025-09-26 02:36:50.535 UTC]  Incoming: [0.073 Gbps] (9.182 MB/s)  total=799.113216 MB
+[2025-09-26 02:36:52.537 UTC]  Incoming: [0.073 Gbps] (9.182 MB/s)  total=817.496064 MB
+[2025-09-26 02:36:54.540 UTC]  Incoming: [0.069 Gbps] (8.642 MB/s)  total=834.797568 MB
+[2025-09-26 02:36:56.541 UTC]  Incoming: [0.073 Gbps] (9.183 MB/s)  total=853.180416 MB
+[2025-09-26 02:36:58.544 UTC]  Incoming: [0.069 Gbps] (8.642 MB/s)  total=870.48192 MB
+[2025-09-26 02:37:00.546 UTC]  Incoming: [0.073 Gbps] (9.182 MB/s)  total=888.864768 MB
+[2025-09-26 02:37:02.548 UTC]  Incoming: [0.073 Gbps] (9.182 MB/s)  total=907.247616 MB
+[2025-09-26 02:37:04.550 UTC]  Incoming: [0.069 Gbps] (8.642 MB/s)  total=924.54912 MB
+[2025-09-26 02:37:06.552 UTC]  Incoming: [0.073 Gbps] (9.182 MB/s)  total=942.931968 MB
+[2025-09-26 02:37:08.554 UTC]  Incoming: [0.069 Gbps] (8.642 MB/s)  total=960.233472 MB
+[2025-09-26 02:37:10.556 UTC]  Incoming: [0.073 Gbps] (9.182 MB/s)  total=978.61632 MB
+[2025-09-26 02:37:12.558 UTC]  Incoming: [0.073 Gbps] (9.183 MB/s)  total=996.999168 MB
+[2025-09-26 02:37:14.560 UTC]  Incoming: [0.069 Gbps] (8.642 MB/s)  total=1014.300672 MB
+[2025-09-26 02:37:16.562 UTC]  Incoming: [0.073 Gbps] (9.182 MB/s)  total=1032.68352 MB
+[2025-09-26 02:37:18.564 UTC]  Incoming: [0.073 Gbps] (9.182 MB/s)  total=1051.066368 MB
+[2025-09-26 02:37:20.566 UTC]  Incoming: [0.069 Gbps] (8.641 MB/s)  total=1068.367872 MB
+[2025-09-26 02:37:22.568 UTC]  Incoming: [0.052 Gbps] (6.482 MB/s)  total=1081.344 MB
